requirements.txt
mypy
sqlalchemy
dataclasses-json
google-api-python-client
google-auth-httplib2
google-auth-oauthlib